split recipe context menu into wrapper and content
